+++ CHANGELOG.md
@@ -14,6 +14,10 @@ maddy uses [semver versioning](https://semver.org/).
 
 ## Upcoming
 
+* ...
+
+## version 1.5.0 2025-04-21
+
 * ![**ADDED**](https://img.shields.io/badge/-ADDED-%23099) Correctly parse links with title text, i.e. `[link](http://example.com "example")`.
 * ![**FIXED**](https://img.shields.io/badge/-FIXED-%23090) Do not create invalid URLs from links with spaces, i.e. `[link](/ABC/some file)`.
 * ![**FIXED**](https://img.shields.io/badge/-FIXED-%23090) Do not create invalid HTML from links with quotes, i.e. `[link](/ABC/some"file)`.